requirements.txt
astropy
astroplan
sphinx
sphinxcontrib.napoleon
sphinx_rtd_theme
pyyaml
